Admin and User Sections complete with security - needs tweaking

+++ app/Http/Controllers/AdminUsersController.php
@@ -167,7 +167,12 @@ class AdminUsersController extends Controller
 
         $user = User::findOrFail($id);
 
-        unlink(public_path() . $user->photo->file);
+        if($file = $this->file('photo_id')){
+
+            unlink(public_path() . $user->photo->file);
+
+        }
+//        unlink(public_path() . $user->photo->file);
 
         $user->delete();
